pyproject.toml
[project]
name = "sam3-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "decord>=0.6.0",
    "einops>=0.8.2",
    "gdal>=3.12.1",
    "huggingface-hub>=0.36.1",
    "matplotlib>=3.10.8",
    "opencv-python>=4.8,<4.10",
    "pillow>=12.1.0",
    "pycocotools>=2.0.11",
    "rasterio>=1.4.4",
    "sam3",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=4.57.6",
]

[project.scripts]
sam3-example = "sam3_example:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "jupyter>=1.1.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]

[tool.uv.sources]
sam3 = { git = "https://github.com/facebookresearch/sam3.git" }
